requirements.txt
aiohttp==3.6.2
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.10
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
loguru==0.5.3
pylint==2.6.0
pymongo==3.10.1
requests==2.24.0
python-dotenv==0.17.0
Flask-Logging==0.1.3